feat(0X-Guide): add debug testing folder
add base objectFlattener to index
